Moving posix version of arg_split to _process_common.
