+++ IPython/utils/_process_posix.py
@@ -18,13 +18,12 @@ from __future__ import print_function
 # Stdlib
 import subprocess as sp
 import sys
-import shlex
 
 from IPython.external import pexpect
 
 # Our own
 from .autoattr import auto_attr
-from ._process_common import getoutput
+from ._process_common import getoutput, arg_split
 from IPython.utils import text
 from IPython.utils import py3compat
 
@@ -193,29 +192,6 @@ class ProcessHandler(object):
 # programs think they are talking to a tty and produce highly formatted output
 # (ls is a good example) that makes them hard.
 system = ProcessHandler().system
-
-def arg_split(s, posix=False):
-    """Split a command line's arguments in a shell-like manner.
-
-    This is a modified version of the standard library's shlex.split()
-    function, but with a default of posix=False for splitting, so that quotes
-    in inputs are respected."""
-
-    # Unfortunately, python's shlex module is buggy with unicode input:
-    # http://bugs.python.org/issue1170
-    # At least encoding the input when it's unicode seems to help, but there
-    # may be more problems lurking.  Apparently this is fixed in python3.
-    is_unicode = False
-    if (not py3compat.PY3) and isinstance(s, unicode):
-        is_unicode = True
-        s = s.encode('utf-8')
-    lex = shlex.shlex(s, posix=posix)
-    lex.whitespace_split = True
-    tokens = list(lex)
-    if is_unicode:
-        # Convert the tokens back to unicode.
-        tokens = [x.decode('utf-8') for x in tokens]
-    return tokens
